Fix problems with profiling by replacing Float type with CGFloat

+++ BouncyView/BouncyView.swift
@@ -11,11 +11,11 @@ import CoreGraphics
 
 class BouncyView: UIView {
 
-    var sideToCenterDelta: Float = 0.0
+    var sideToCenterDelta: CGFloat = 0.0
     let fillColor = UIColor(red: 0, green: 0.722, blue: 1, alpha: 1) // blue color
 
     override func drawRect(rect: CGRect) {
-        let yOffset: Float = 20.0
+        let yOffset: CGFloat = 20.0
         let width = CGRectGetWidth(rect)
         let height = CGRectGetHeight(rect)
